+++ packages/andromeda-non-fungible-tokens/src/auction.rs
@@ -2,10 +2,11 @@ use andromeda_std::amp::{AndrAddr, Recipient};
 use andromeda_std::common::denom::Asset;
 use andromeda_std::common::expiration::Expiry;
 use andromeda_std::common::{MillisecondsExpiration, OrderBy};
+use andromeda_std::error::ContractError;
 use andromeda_std::{andr_exec, andr_instantiate, andr_query};
 
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::{Addr, Uint128};
+use cosmwasm_std::{ensure, Addr, BlockInfo, MessageInfo, Uint128};
 use cw20::Cw20ReceiveMsg;
 use cw721::{Cw721ReceiveMsg, Expiration};
 
@@ -25,6 +26,10 @@ pub enum ExecuteMsg {
     /// Places a bid on the current auction for the given token_id. The previous largest bid gets
     /// automatically sent back to the bidder when they are outbid.
     PlaceBid {
+        token_id: String,
+        token_address: String,
+    },
+    BuyNow {
         token_id: String,
         token_address: String,
     },
@@ -69,6 +74,7 @@ pub enum Cw721HookMsg {
         /// Duration in milliseconds
         end_time: Expiry,
         coin_denom: Asset,
+        buy_now_price: Option<Uint128>,
         min_bid: Option<Uint128>,
         min_raise: Option<Uint128>,
         whitelist: Option<Vec<Addr>>,
@@ -78,6 +84,10 @@ pub enum Cw721HookMsg {
 #[cw_serde]
 pub enum Cw20HookMsg {
     PlaceBid {
+        token_id: String,
+        token_address: String,
+    },
+    BuyNow {
         token_id: String,
         token_address: String,
     },
@@ -191,6 +201,7 @@ pub struct TokenAuctionState {
     pub end_time: Expiration,
     pub high_bidder_addr: Addr,
     pub high_bidder_amount: Uint128,
+    pub buy_now_price: Option<Uint128>,
     pub coin_denom: String,
     pub auction_id: Uint128,
     pub whitelist: Option<Vec<Addr>>,
@@ -200,6 +211,7 @@ pub struct TokenAuctionState {
     pub token_id: String,
     pub token_address: String,
     pub is_cancelled: bool,
+    pub is_bought: bool,
     pub uses_cw20: bool,
     pub recipient: Option<Recipient>,
 }
@@ -209,6 +221,49 @@ pub struct Bid {
     pub bidder: String,
     pub amount: Uint128,
     pub timestamp: MillisecondsExpiration,
+}
+
+/// Checks against auctions that are: cancelled, not started, already bought, and ended.
+/// Also checks for token owner bidding and funds being exactly of one denomination
+pub fn validate_auction(
+    token_auction_state: TokenAuctionState,
+    info: MessageInfo,
+    block: &BlockInfo,
+) -> Result<(), ContractError> {
+    ensure!(
+        !token_auction_state.is_cancelled,
+        ContractError::AuctionCancelled {}
+    );
+
+    ensure!(
+        !token_auction_state.is_bought,
+        ContractError::AuctionBought {}
+    );
+
+    ensure!(
+        token_auction_state.start_time.is_expired(block),
+        ContractError::AuctionNotStarted {}
+    );
+    ensure!(
+        !token_auction_state.end_time.is_expired(block),
+        ContractError::AuctionEnded {}
+    );
+
+    ensure!(
+        token_auction_state.owner != info.sender.clone().into_string(),
+        ContractError::TokenOwnerCannotBid {}
+    );
+
+    if !token_auction_state.uses_cw20 {
+        ensure!(
+            info.funds.len() == 1,
+            ContractError::InvalidFunds {
+                msg: "One coin should be sent.".to_string(),
+            }
+        );
+    }
+
+    Ok(())
 }
 
 #[cw_serde]