+++ packages/std/src/error.rs
@@ -135,6 +135,9 @@ pub enum ContractError {
     #[error("NoTokens")]
     NoTokens {},
 
+    #[error("NoBuyNowOption")]
+    NoBuyNowOption {},
+
     #[error("UnrecognisedReplyId")]
     UnrecognisedReplyId {},
 
@@ -315,6 +318,9 @@ pub enum ContractError {
     #[error("AuctionEnded")]
     AuctionEnded {},
 
+    #[error("AuctionBought")]
+    AuctionBought {},
+
     #[error("CampaignNotStarted")]
     CampaignNotStarted {},
 
@@ -392,6 +398,9 @@ pub enum ContractError {
 
     #[error("InvalidADOVersion: {msg:?}")]
     InvalidADOVersion { msg: Option<String> },
+
+    #[error("InvalidMinBid: {msg:?}")]
+    InvalidMinBid { msg: Option<String> },
 
     #[error("InvalidCodeID: {msg:?}")]
     InvalidCodeID { msg: Option<String> },